requirements.txt
requests==2.32.3
markdown==3.7
beautifulsoup4==4.12.3
qdrant-client==1.7.3
chromadb==0.5.12
python-frontmatter==1.0.1
PyYAML==6.0.2
textstat==0.7.4
numpy==1.26.4
watchdog==4.0.1
pdfminer.six==20231228
sentence-transformers==3.0.1
fastapi==0.111.1
uvicorn[standard]==0.30.3